Added the ability to leave the generated temporary file for debuging tests.

+++ volttron/tests/base.py
@@ -19,7 +19,7 @@ try:
 except ImportError:
     auth = None
     certs = None
-    
+
 
 # from volttron.platform.control import (CTL_STATUS,
 #                                        CTL_INSTALL,
@@ -90,58 +90,60 @@ print VSTART
 
 class BasePlatformTest(unittest.TestCase):
 
-    def setUp(self): 
+    cleanup_tempdir = True
+
+    def setUp(self):
         self.tmpdir = tempfile.mkdtemp()
         self.wheelhouse = '/'.join((self.tmpdir, 'wheelhouse'))
         os.makedirs(self.wheelhouse)
         os.environ['VOLTTRON_HOME'] = self.tmpdir
-        
+
         opts = type('Options', (), {'verify_agents': False, 'volttron_home': self.tmpdir})()
         self.test_aip = aip.AIPplatform(opts)
         self.test_aip.setup()
-        
-        
+
+
     def setup_connector(self):
         path = os.path.expandvars('$VOLTTRON_HOME/run/control')
-        
+
         tries = 0
         max_tries = 5
         while(not os.path.exists(path) and tries < max_tries):
             time.sleep(5)
             tries += 1
-            
+
         self.conn= server.ControlConnector(path)
-    
+
     def startup_platform(self, platform_config, volttron_home=None, use_twistd = True, mode=UNRESTRICTED):
-        
-        
-        
+
+
+
         try:
             config = json.loads(open(platform_config, 'r').read())
         except Exception as e:
             sys.stderr.write (str(e))
             self.fail("Could not load configuration file for tests {}".format(e))
-        
+
 #         self.tmpdir = tempfile.mkdtemp()
         config['tmpdir'] = self.tmpdir
-        
+
         pconfig = os.path.join(self.tmpdir, TMP_PLATFORM_CONFIG_FILENAME)
-        
+
         self.mode = mode
-        
+
         self.assertIn(self.mode, MODES, 'Invalid platform mode set: '+str(mode))
-        
+
         if self.mode == UNRESTRICTED:
             with closing(open(pconfig, 'w')) as cfg:
                 cfg.write(PLATFORM_CONFIG_UNRESTRICTED.format(**config))
         elif self.mode == RESTRICTED:
             with closing(open(pconfig, 'w')) as cfg:
                 cfg.write(PLATFORM_CONFIG_RESTRICTED.format(**config))
-            
-            
+
+
 #                 self.create_certs()
         else:
-            
+
             self.fail("Platform mode not implemented: "+str(mode))
 
         tconfig = os.path.join(self.tmpdir, TMP_SMAP_CONFIG_FILENAME)
@@ -154,11 +156,11 @@ class BasePlatformTest(unittest.TestCase):
         pparams = [VSTART, "-c", pconfig, "-vv", "-l", lfile]
         print pparams
         self.p_process = subprocess.Popen(pparams)
-        
+
         self.setup_connector()
         if self.mode == RESTRICTED:
             self.conn.call.create_cgroups()
-        
+
         self.use_twistd = use_twistd
         if self.use_twistd:
             with closing(open(tconfig, 'w')) as cfg:
@@ -169,23 +171,23 @@ class BasePlatformTest(unittest.TestCase):
             time.sleep(5)
         #self.t_process = subprocess.Popen(["twistd", "-n", "smap", "test-smap.ini"])
 
-        
+
 
     def fillout_file(self, filename, template, config_file):
-        
+
         try:
             config = json.loads(open(config_file, 'r').read())
         except Exception as e:
             sys.stderr.write (str(e))
             self.fail("Could not load configuration file for tests")
-        
+
 #         self.tmpdir = tempfile.mkdtemp()
         config['tmpdir'] = self.tmpdir
-        
+
         outfile = os.path.join(self.tmpdir, filename)
         with closing(open(outfile, 'w')) as cfg:
             cfg.write(template.format(**config))
-            
+
         return outfile
 
 
@@ -198,34 +200,34 @@ class BasePlatformTest(unittest.TestCase):
         try:
             basepackage = os.path.join(self.tmpdir,distdir)
             shutil.copytree(os.path.abspath(distdir), basepackage)
-            
+
             os.chdir(basepackage)
             sys.argv = ['', 'bdist_wheel']
             exec(compile(open('setup.py').read(), 'setup.py', 'exec'))
-     
+
             wheel_name = os.listdir('./dist')[0]
-     
+
             wheel_file_and_path = os.path.join(os.path.abspath('./dist'), wheel_name)
         finally:
             os.chdir(pwd)
-            
+
         return wheel_file_and_path
-    
+
     def direct_build_agentpackage(self, agent_dir):
         wheel_path = packaging.create_package(os.path.join(rel_path, agent_dir), self.wheelhouse)
-            
+
         return wheel_path
 
     def direct_configure_agentpackage(self, agent_wheel, config_file):
         packaging.add_files_to_package(agent_wheel, {'config_file':os.path.join(rel_path, config_file)})
-            
+
 
 
     def direct_buid_install_agent(self, agent_dir, config_file):
         agent_wheel = self.direct_build_agentpackage(agent_dir)
         self.direct_configure_agentpackage(agent_wheel, config_file)
         self.assertIsNotNone(agent_wheel,"Agent wheel was not built")
-        
+
         uuid = self.test_aip.install_agent(agent_wheel)
         #aip volttron_home, verify_agents
         return uuid
@@ -236,33 +238,35 @@ class BasePlatformTest(unittest.TestCase):
 
     def direct_build_install_run_agent(self, agent_dir, config_file):
         agent_uuid = self.direct_buid_install_agent(agent_dir, config_file)
-        self.direct_start_agent(agent_uuid)  
+        self.direct_start_agent(agent_uuid)
         return agent_uuid
-            
+
     def direct_start_agent(self, agent_uuid):
-        
+
         self.conn.call.start_agent(agent_uuid)
         time.sleep(3)
-        
+
         status = self.conn.call.status_agents()
 #         self.test_aip.status_agents()
-        
+
 #         status = self.conn.call.status_agents()
 #         self.assertEquals(len(status[0]), 4, 'Unexpected status message')
         status_uuid = status[0][0]
         self.assertEquals(status_uuid, agent_uuid, "Agent status shows error")
-        
+
         self.assertEquals(len(status[0][2]), 2, 'Unexpected agent status message')
         status_agent_status = status[0][2][1]
+        print("AGENT STATUS: ", isinstance(status_agent_status, int))
+        print("AGENT STATUS: ", isinstance(status_agent_status, str))
         self.assertNotIsInstance(status_agent_status, int, "Agent did not start successfully")
 #         self.assertIn("running",status_agent_status, "Agent status shows error")
         #print status
-        
+
     def direct_stop_agent(self, agent_uuid):
         result = self.conn.call.stop_agent(agent_uuid)
         print result
 
-            
+
     def check_default_dir(self, dir_to_check):
         default_dir = os.path.expanduser("~/.volttron/agents")
         self.assertNotIn(default_dir, dir_to_check, "Platform is using defaults")
@@ -282,17 +286,19 @@ class BasePlatformTest(unittest.TestCase):
             self.t_process.wait()
         elif self.use_twistd:
             print "twistd process was null"
-        if self.tmpdir != None:
-            shutil.rmtree(self.tmpdir, True)
-    
+        if  self.cleanup_tempdir:
+            if self.tmpdir != None:
+                shutil.rmtree(self.tmpdir, True)
+
     def tearDown(self):
         try:
             self.shutdown_platform()
         except Exception as e:
             sys.stderr.write( str(e))
         finally:
-            shutil.rmtree(self.tmpdir,True)
-            
+            if self.cleanup_tempdir:
+                shutil.rmtree(self.tmpdir,True)
+
 
 def mergetree(src, dst, symlinks=False, ignore=None):
     if not os.path.exists(dst):
@@ -313,14 +319,14 @@ def mergetree(src, dst, symlinks=False, ignore=None):
 #         results = subprocess.check_output([VCTRL,CTL_STATUS])
 #         print results
 #         self.assertIn('running', results, "Agent was not started")
-#         
+#
 
 #     def build_and_install_agent(self, agent_dir):
 #         agent_wheel = self.build_agentpackage(agent_dir)
 #         self.assertIsNotNone(agent_wheel,"Agent wheel was not built")
-#         
+#
 #         sys.stderr = std_err = StringIO()
-#         
+#
 #         results = subprocess.check_output([VCTRL,CTL_INSTALL,agent_wheel])
 #         print ("results: "+results)
 #         if self.mode == UNRESTRICTED or self.mode == RESOURCE_CHECK_ONLY:
@@ -328,5 +334,5 @@ def mergetree(src, dst, symlinks=False, ignore=None):
 #             proc_out = std_err.getvalue().split(':')
 #             print (proc_out)
 #             self.check_default_dir(proc_out[1].strip())
-#         elif self.mode == RESTRICTED or self.mode == VERIFY_ONLY:    
+#         elif self.mode == RESTRICTED or self.mode == VERIFY_ONLY:
 #             self.assertTrue(results.startswith('Unpacking to: '))
